# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IdeaGPT API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ideas/{idea_id}/evaluations": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Trigger Evaluation",
        "description": "Creates and executes an AI or deterministic evaluation job for an idea.",
        "operationId": "trigger_evaluation_api_v1_ideas__idea_id__evaluations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Idea Evaluations",
        "description": "Lists all evaluation jobs for a specific idea.",
        "operationId": "get_idea_evaluations_api_v1_ideas__idea_id__evaluations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationResponse"
                  },
                  "title": "Response Get Idea Evaluations Api V1 Ideas  Idea Id  Evaluations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation",
        "description": "Retrieves evaluation status and report payload.",
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "evaluations"
        ],
        "summary": "Delete Evaluation",
        "description": "Deletes an evaluation record.",
        "operationId": "delete_evaluation_api_v1_evaluations__evaluation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/evaluations": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Project Evaluations",
        "description": "Lists all evaluation jobs for all ideas in a specific project.",
        "operationId": "get_project_evaluations_api_v1_projects__project_id__evaluations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationResponse"
                  },
                  "title": "Response Get Project Evaluations Api V1 Projects  Project Id  Evaluations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/retry": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Retry Evaluation",
        "description": "Retries a FAILED or CANCELLED evaluation job.",
        "operationId": "retry_evaluation_api_v1_evaluations__evaluation_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/cancel": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Cancel Evaluation",
        "description": "Cancels an active (PENDING or RUNNING) evaluation job.",
        "operationId": "cancel_evaluation_api_v1_evaluations__evaluation_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/run": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Run Evaluation Pipeline",
        "description": "Triggers/runs the evaluation pipeline via EvaluationCoordinator.",
        "operationId": "run_evaluation_pipeline_api_v1_evaluations__evaluation_id__run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/insights": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation Insights",
        "description": "Retrieves granular structured insights (SWOT, Feasibility, etc.)",
        "operationId": "get_evaluation_insights_api_v1_evaluations__evaluation_id__insights_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/scores": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation Scores",
        "description": "Retrieves multi-dimensional analysis scores.",
        "operationId": "get_evaluation_scores_api_v1_evaluations__evaluation_id__scores_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/history": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation History",
        "description": "Retrieves historical lifecycle audit events for an evaluation.",
        "operationId": "get_evaluation_history_api_v1_evaluations__evaluation_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationHistoryResponse"
                  },
                  "title": "Response Get Evaluation History Api V1 Evaluations  Evaluation Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/compare": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Compare Ideas",
        "description": "Compares 2-5 user-owned ideas using real persisted data and evaluation metrics.\nEnforces user ownership isolation across all selected ideas.",
        "operationId": "compare_ideas_api_v1_evaluations_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/projects/{project_id}/comparisons": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Project Comparisons",
        "description": "Retrieves comparison matrix for multiple ideas/evaluations.",
        "operationId": "get_project_comparisons_api_v1_projects__project_id__comparisons_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "evaluation_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evaluation Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/export": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation Export",
        "description": "RESTful GET export for evaluation payload as JSON, Markdown, or PDF Printable HTML.",
        "operationId": "get_evaluation_export_api_v1_evaluations__evaluation_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|markdown|md|pdf|html)$",
              "description": "Export format: json, markdown, pdf, or html",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format: json, markdown, pdf, or html"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/diff": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Semantic version diffing between two evaluation reports",
        "operationId": "diff_evaluations_api_v1_evaluations_diff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exports/json": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Export Evaluation Json",
        "description": "Exports evaluation payload as raw JSON.",
        "operationId": "export_evaluation_json_api_v1_exports_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_export_evaluation_json_api_v1_exports_json_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationCreate": {
        "properties": {
          "evaluation_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Type",
            "default": "startup_evaluation"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "title": "EvaluationCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdeaCompareRequest": {
        "properties": {
          "idea_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Idea Ids"
          }
        },
        "type": "object",
        "required": [
          "idea_ids"
        ],
        "title": "IdeaCompareRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}